requirements.txt
asgiref==3.8.1
dj-database-url==2.3.0
Django==5.0.7
django-cors-headers==4.7.0
django-mssql-backend==2.8.1
django-rest==0.8.7
gunicorn==23.0.0
mysqlclient==2.2.7
packaging==24.2
pillow==10.4.0
psycopg2==2.9.10
PyJWT==2.10.1
pyodbc==5.2.0
six==1.16.0
source==1.2.0
sqlparse==0.5.1
typing_extensions==4.13.0
tzdata==2024.1
